AF-1028 : [Project Oriented] Standalone perspectives are broken
