+++ kie-wb-parent/kie-wb-webapp/src/main/java/org/kie/workbench/client/perspectives/DroolsAuthoringPerspective.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 Red Hat, Inc. and/or its affiliates.
+ * Copyright Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,31 +15,36 @@
  */
 package org.kie.workbench.client.perspectives;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
+import com.google.gwt.user.client.Window;
+import org.jboss.errai.common.client.api.Caller;
+import org.jboss.errai.common.client.api.RemoteCallback;
 import org.kie.workbench.client.resources.i18n.AppConstants;
-import org.kie.workbench.common.screens.examples.client.wizard.ExamplesWizard;
-import org.kie.workbench.common.screens.examples.service.ExamplesService;
-import org.kie.workbench.common.screens.projecteditor.client.menu.ProjectMenu;
-import org.kie.workbench.common.services.shared.preferences.ApplicationPreferences;
-import org.kie.workbench.common.widgets.client.handlers.NewResourcePresenter;
 import org.kie.workbench.common.widgets.client.handlers.NewResourcesMenu;
-import org.kie.workbench.common.widgets.client.menu.RepositoryMenu;
 import org.kie.workbench.common.workbench.client.PerspectiveIds;
 import org.kie.workbench.common.workbench.client.docks.AuthoringWorkbenchDocks;
+import org.uberfire.backend.vfs.Path;
+import org.uberfire.backend.vfs.VFSService;
 import org.uberfire.client.annotations.Perspective;
 import org.uberfire.client.annotations.WorkbenchMenu;
 import org.uberfire.client.annotations.WorkbenchPerspective;
 import org.uberfire.client.mvp.PlaceManager;
+import org.uberfire.client.workbench.PanelManager;
 import org.uberfire.client.workbench.panels.impl.MultiListWorkbenchPanelPresenter;
-import org.uberfire.mvp.Command;
+import org.uberfire.lifecycle.OnOpen;
+import org.uberfire.mvp.PlaceRequest;
 import org.uberfire.mvp.impl.DefaultPlaceRequest;
+import org.uberfire.workbench.model.PanelDefinition;
+import org.uberfire.workbench.model.PartDefinition;
 import org.uberfire.workbench.model.PerspectiveDefinition;
 import org.uberfire.workbench.model.impl.PerspectiveDefinitionImpl;
 import org.uberfire.workbench.model.menu.MenuFactory;
-import org.uberfire.workbench.model.menu.MenuPosition;
 import org.uberfire.workbench.model.menu.Menus;
 
 @ApplicationScoped
@@ -49,105 +54,113 @@ public class DroolsAuthoringPerspective {
     private AppConstants constants = AppConstants.INSTANCE;
 
     @Inject
-    private NewResourcePresenter newResourcePresenter;
-
-    @Inject
     private NewResourcesMenu newResourcesMenu;
-
-    @Inject
-    private ProjectMenu projectMenu;
 
     @Inject
     private PlaceManager placeManager;
 
     @Inject
-    private RepositoryMenu repositoryMenu;
+    private PanelManager panelManager;
+
+    @Inject
+    private Caller<VFSService> vfsServices;
 
     @Inject
     private AuthoringWorkbenchDocks docks;
 
-    @Inject
-    private ExamplesWizard wizard;
+    private String explorerMode;
+    private String projectPathString;
+    private boolean projectEditorDisableBuild;
+
+    private final List<PlaceRequest> placesToClose = new ArrayList<PlaceRequest>();
 
     @PostConstruct
-    public void setup() {
-        docks.setup( PerspectiveIds.AUTHORING, new DefaultPlaceRequest( "org.kie.guvnor.explorer" ) );
+    public void init() {
+        explorerMode = ( ( Window.Location.getParameterMap().containsKey( "explorer_mode" ) ) ? Window.Location.getParameterMap().get( "explorer_mode" ).get( 0 ) : "" ).trim();
+        projectPathString = ( ( ( Window.Location.getParameterMap().containsKey( "path" ) ) ? Window.Location.getParameterMap().get( "path" ).get( 0 ) : "" ) ).trim();
+        projectEditorDisableBuild = Window.Location.getParameterMap().containsKey( "no_build" );
+
+        final PlaceRequest placeRequest = generateProjectExplorerPlaceRequest();
+
+        docks.setup(PerspectiveIds.AUTHORING, placeRequest);
+
+    }
+
+    private PlaceRequest generateProjectExplorerPlaceRequest() {
+        final PlaceRequest placeRequest = new DefaultPlaceRequest( "org.kie.guvnor.explorer" );
+        if ( !explorerMode.isEmpty() ) {
+            placeRequest.addParameter( "mode",
+                    explorerMode );
+        }
+        if ( !projectPathString.isEmpty() ) {
+            placeRequest.addParameter( "init_path",
+                    projectPathString );
+        }
+        if ( projectEditorDisableBuild ) {
+            placeRequest.addParameter( "no_build",
+                    "true" );
+        }
+
+        placeRequest.addParameter( "no_context",
+                "true" );
+        return placeRequest;
     }
 
     @Perspective
     public PerspectiveDefinition getPerspective() {
-        PerspectiveDefinitionImpl perspective = new PerspectiveDefinitionImpl( MultiListWorkbenchPanelPresenter.class.getName() );
-        perspective.setName( constants.Project_Authoring() );
+        final PerspectiveDefinitionImpl perspective = new PerspectiveDefinitionImpl( MultiListWorkbenchPanelPresenter.class.getName() );
+        perspective.setName(constants.Project_Authoring());
 
         return perspective;
     }
 
-    @WorkbenchMenu
-    public Menus getMenus() {
-        if ( !ApplicationPreferences.isProductized() && ApplicationPreferences.getBooleanPref( ExamplesService.EXAMPLES_SYSTEM_PROPERTY ) ) {
-            return buildMenuBarWithExamples();
+    @OnOpen
+    public void onOpen() {
+        placesToClose.clear();
+        if ( !projectPathString.isEmpty() ) {
+            vfsServices.call( new RemoteCallback<Boolean>() {
+                @Override
+                public void callback( Boolean isRegularFile ) {
+                    if ( isRegularFile ) {
+                        vfsServices.call( new RemoteCallback<Path>() {
+                            @Override
+                            public void callback( Path path ) {
+                                placeManager.goTo( path );
+                            }
+                        } ).get( projectPathString );
+                    }
+                }
+            } ).isRegularFile( projectPathString );
+        }
+        if ( panelManager.getRoot() != null ) {
+            process( panelManager.getRoot().getParts() );
+            process( panelManager.getRoot().getChildren() );
 
-        } else {
-            return buildMenuBarWithoutExamples();
+            for ( final PlaceRequest placeRequest : placesToClose ) {
+                placeManager.forceClosePlace( placeRequest );
+            }
         }
     }
 
-    private Menus buildMenuBarWithExamples() {
-        return MenuFactory
-                .newTopLevelMenu( constants.Examples() )
-                .respondsWith( new Command() {
-                    @Override
-                    public void execute() {
-                        wizard.start();
-                    }
-                } )
-                .endMenu()
-                .newTopLevelMenu( constants.newItem() )
-                .withItems( newResourcesMenu.getMenuItems() )
-                .endMenu()
-                .newTopLevelMenu( AppConstants.INSTANCE.Repository() )
-                .withItems( repositoryMenu.getMenuItems() )
-                .endMenu()
-                .newTopLevelMenu( constants.assetSearch() ).position( MenuPosition.RIGHT ).respondsWith( new Command() {
-                    @Override
-                    public void execute() {
-                        placeManager.goTo( "FindForm" );
-                    }
-                } )
-                .endMenu()
-                .newTopLevelMenu( constants.Messages() ).position( MenuPosition.RIGHT ).respondsWith( new Command() {
-                    @Override
-                    public void execute() {
-                        placeManager.goTo( "org.kie.workbench.common.screens.messageconsole.MessageConsole" );
-                    }
-                } )
-                .endMenu()
-                .build();
+    private void process( final List<PanelDefinition> children ) {
+        for ( final PanelDefinition child : children ) {
+            process( child.getParts() );
+            process( child.getChildren() );
+        }
     }
 
-    private Menus buildMenuBarWithoutExamples() {
-        return MenuFactory
-                .newTopLevelMenu( constants.newItem() )
-                .withItems( newResourcesMenu.getMenuItems() )
-                .endMenu()
-                .newTopLevelMenu( AppConstants.INSTANCE.Repository() )
-                .withItems( repositoryMenu.getMenuItems() )
-                .endMenu()
-                .newTopLevelMenu( constants.assetSearch() ).position( MenuPosition.RIGHT ).respondsWith( new Command() {
-                    @Override
-                    public void execute() {
-                        placeManager.goTo( "FindForm" );
-                    }
-                } )
-                .endMenu()
-                .newTopLevelMenu( constants.Messages() ).position( MenuPosition.RIGHT ).respondsWith( new Command() {
-                    @Override
-                    public void execute() {
-                        placeManager.goTo( "org.kie.workbench.common.screens.messageconsole.MessageConsole" );
-                    }
-                } )
-                .endMenu()
-                .build();
+    private void process( final Collection<PartDefinition> parts ) {
+        for ( final PartDefinition partDefinition : parts ) {
+            if ( !partDefinition.getPlace().getIdentifier().equals( "org.kie.guvnor.explorer" ) ) {
+                placesToClose.add( partDefinition.getPlace() );
+            }
+        }
     }
 
+    @WorkbenchMenu
+    public Menus getMenus() {
+        return MenuFactory.newTopLevelMenu( constants.newItem() )
+                .withItems( newResourcesMenu.getMenuItemsWithoutProject() ).endMenu()
+                .build();
+    }
 }
